project/pyproject.toml
    [project]
    name = "project"
    version = "0.1.0"
    description = ""
    authors = [
        {name = "Pablo Chantada Saborido",email = "pablo.chantada@diverger.ai"}
    ]
    readme = "README.md"
    requires-python = ">=3.10,<3.15"
    dependencies = [
        "pandas (>=2.3.3,<3.0.0)",
        "pydantic (>=2.11.10,<3.0.0)",
        "jsonschema (>=4.25.1,<5.0.0)",
        "python-dotenv (>=1.1.1,<2.0.0)",
        "jinja2 (>=3.1.6,<4.0.0)",
        "fastapi[all] (>=0.118.0,<0.119.0)",
        "uvicorn[standart] (>=0.37.0,<0.38.0)",
        "google-genai (>=1.41.0,<2.0.0)",
        "langchain (>=0.3.27,<0.4.0)",
        "scikit-learn (>=1.7.2,<2.0.0)",
        "pytest (>=8.4.2,<9.0.0)",
        "google-adk (>=1.16.0,<2.0.0)",
    ]

    [tool.poetry]
    packages = [{include = "project", from = "src"}]

    [build-system]
    requires = ["poetry-core>=2.0.0,<3.0.0"]
    build-backend = "poetry.core.masonry.api"
